Refactor method to extract route and params
Also updated tests
